requirements.txt
flask==2.3.3
flask-restx==1.1.0
flask_cors
pymongo
werkzeug==3.0.6
certifi
bcrypt==4.1.2
pyjwt==2.8.0